Configure client and server within project
Use the client to handle any navigation requests from ReactRouter

Use the server to handle any fresh requests for direct pages, and then
forward the request to the client for hydration.

Rebuild the server at '/server/index.js' on every change to a file
within the '/server/' directory.

Rebuild the client at '/build/...somePage' on every change to a file
within the '/server/' or '/client/' directory.